Move functions DB entities to functions package
This is needed for firebase version upgrade.

Common cannot depend on firebase-admin
