+++ packages/common/src/types.ts
@@ -1,4 +1,3 @@
-import firebase from "firebase";
 // Hospital has available machines
 // Admin opens new slots in app
 // Donors sign up in slots - if they are allowed to
@@ -24,24 +23,6 @@ export type DonorNotificationSettings = {
   disableEmailNotifications: boolean;
 };
 
-export type DbDonor = {
-  id: string; // UUID generated by firebase
-  email: string;
-  firstName: string;
-  lastName: string;
-  birthDate: string; // YYYY-MM-DD
-  phone: string;
-  bloodType: BloodType;
-  groupId: string;
-  notificationSettings?: DonorNotificationSettings;
-
-  // Fields changing after booking a donation
-  lastBookingTime?: firebase.firestore.Timestamp;
-  lastBookedHospital?: Hospital; // The hospital of the last booked donation of the donor
-
-  testUser: boolean;
-};
-
 export type Donor = {
   id: string;
   email: string;
@@ -61,22 +42,10 @@ export enum CoordinatorRole {
   GROUP_COORDINATOR = "GROUP_COORDINATOR",
 }
 
-export type DbCoordinator = {
-  id: string; // UUID generated by firebase
-  role: CoordinatorRole;
-  hospitals?: Hospital[];
-};
-
 export type Coordinator = {
   role: CoordinatorRole;
   activeHospitalsForCoordinator: Hospital[];
   name: string | undefined;
-};
-
-export type DbGroup = {
-  id: string; // UUID generated by firebase
-  name: string;
-  coordinatorId: string;
 };
 
 export enum Hospital {
@@ -100,26 +69,6 @@ export type BookedAppointment = {
   hospital: Hospital;
   donorId: string;
   bookingTimeMillis: number;
-};
-
-export type DbAppointment = {
-  id?: string;
-
-  // Pre-donor
-  creationTime: firebase.firestore.Timestamp;
-  creatorUserId: string;
-  donationStartTime: firebase.firestore.Timestamp;
-  hospital: Hospital;
-
-  // added donor
-  donorId: string; // Empty string means appointment is available
-  bookingTime?: firebase.firestore.Timestamp;
-  confirmationTime?: firebase.firestore.Timestamp; // Time donor confirmed they will come
-  /* Represents changes to booking status **/
-  lastChangeTime?: firebase.firestore.Timestamp;
-  lastChangeType?: BookingChange;
-  donationDoneTimeMillis?: firebase.firestore.Timestamp;
-  status?: AppointmentStatus; // TODO: remove ? once db migration happens
 };
 
 export type AvailableAppointment = {
