+++ packages/functions/src/donor/BookAppointmentHandler.ts
@@ -8,13 +8,13 @@ import * as admin from "firebase-admin";
 import {
   AppointmentStatus,
   BookingChange,
-  DbDonor,
   FunctionsApi,
   Hospital,
 } from "@zm-blood-components/common";
 import { dbAppointmentToBookedAppointmentApiEntry } from "../utils/ApiEntriesConversionUtils";
 import { notifyOnAppointmentBooked } from "../notifications/BookAppointmentNotifier";
 import { BookAppointmentStatus } from "../../../common/src/functions-api";
+import { DbDonor } from "../function-types";
 
 const WEEKS_BUFFER = 0;
